Update Blog Title
Not sure where the AudioKit blog should go? It's not a company. And, the majority of the iOS Developers we've interviewed don't use AudioKit in their apps. (There's something in there for everyone- tips on getting apps featured, marketing tips, tips for first-time developers). Looking for your guidance on this one. 🙏